Removing $err property
Since the errors are handled by Error() method and stored in the $error_info property, there is no reason to duplicate with the $err property. This adds unnecessary memory usage. Plus, this was only used once in the script, inside ParseXMLResponse, which also already used the set_error_info for storing the error details. 