Fixes an issue with cookies from the last change made.
When both Guzzle and cURL are told to handle cookies, they trip over each other a bit and cause duplicate Cookie headers to be given in requests.  At least the Interealty server seems to have issues with these.  This change tells Guzzle to stop worrying about cookies and leaves that heavy lifting to cURL to do.  If a different HTTP library is used or if Guzzle starts implementing it's own Digest authentication, we should be able to change back to this.  This may cause some issues for Guzzle middleware that wants to look at and manipulate cookies, but that seems like a pretty limited use case so this change seems fairly safe.