Fix issue #172 with tempnam() throwing an Exception.
I have a Homestead server setup on my Windows PC for development and so my temp directory is not "/tmp" and Laravel throws an ErrorException. The `sys_get_temp_dir()` function returns the default path PHP stores temporary files in. So it will still resolve to `"/tmp"` on Linux, but for Windows it will resolve to the correct temp directory.